NoStim improvements for live coding
Allows for more flexible use of the "trial type" column because that information is not required for selecting stimuli in the no-stim version. For VoE designs, the trial type column can now be arbitrary text, controlled either by the trialOrder array or by a random presentation file.

Also, the "NEXT TRIAL" text has been fixed and made optional using a new control boolean, "blindPres". Allows for optional blinding of the coder to trial type.