Fixed an issue with stimName
The root of all this is that at one point counters was cycling through
just something the length of the number of different stimuli, and now
it’s cycling through something that’s the length of the number of
repetitions of a trial type. Previously it could break if you had more
repetitions than 2x the number of movies. Now it doesn’t.