Modularization and new tests
Two new functions: “redoSetup” and “jumpToTest”.
redoSetup is a helper function for redoTrial that updates the trial
number, picks the right stimuli, updates all the relevant counters,
etc., and then calls redoTrial to update the data
jumpToTest is a modularization of the process of going from the middle
of a hab block to the first post-hab test. The function adds something
the old version didn’t have: if you are using hab meta-trials, and
‘Hab’ was not the last part of the meta-trial, it now correctly skips
to the first trial that is not part of the meta-trial.

Testing: The test for PyHabClass has been updated to include tests of
verbose data updating for Redo, and tests of the new study flow
functions.