Insert hab trials + test
Modularized insert hab and fixed it so that it now actually works.
Previously it would crash if you were doing stim presentation because
it would not add a movie to stimDict. Now it actually finds the correct
movie and imports it, which will probably be slow but functional.
Also fixed habituation if using hab meta-trials in which hab is not the
last trial type. Previously would go to the first trial after Hab. Now,
like jumpToTest, goes to the end of the last meta-trial.

Testing: Cannot automatically test media importing effectively, but
there is now a test for inserting both a hab trial and a meta-trial in
terms of study flow.