Saving data and builder tune-up
changed how data are saved to prevent overwriting existing files in
favor of creating a tagged duplicate.

Builder now has the capability to support a functionally unlimited
number of trials, but render up to 40 and provide a counter for every
trial added over 40.

Color/trial type associations are now preserved within a single session
of the builder. They may be reshuffled when the builder is closed and
re-opened, but whatever color a trial is assigned when you create it,
that’s the trial it will have for the rest of that session.