Windows bug fixes
Attempting to address two issues:
1. Converting between Mac and Windows is a little unreliable because
they use different folder markers (\ vs. /). Under some circumstances
this would mean that a study created on Mac would not work on Windows
and vice-versa. That should now be fixed.

2. I got reports of some anomalous save-data behavior on windows where
the write was being delayed somehow and saving blank files, so I’ve
gone to a “safer” python method of using “with” to force python to
write the data file and save it immediately.