New feature + condition warning
NEW FEATURE: When the next trial requires manual input to start (i.e.,
the experiment does not auto-advance into it), PyHab can now be
configured to flash the background of the coder window, in order to get
the experimenter’s attention and let them know they need to start the
next trial.

NEW SAFEGUARD: If a new trial type is added after the condition
settings have been made, the experiment will break. I don’t have a fix
for this (and I don’t want to, because silently filling in your
conditions is NOT what I want PyHab to do), but it now pops up an
explanatory warning message if you add a new trial type after making
conditions.