Updating docs, small change to minOn behavior
Adding better installation instructions and some windows-specific help,
corrected some in-code documentation.

Gaze-contingent trials used to require at least one frame greater than
MinOn in order to allow the trial to end based on MaxOff. This led to
unintuitive behavior if you set minOn to 0, specifically you would
still need to present the gaze-on key at least once even for auto
advance trials. Changed so it is now >= rather than >, meaning that if
you set it to 0, no gaze-on is required.