Builder input safeguards
Added a bunch of code to make it so the builder is now able to catch
illegal inputs (strings that should be numbers, floats that should be
ints, etc.) and prompt the user to fix them. Involved adding a lot of
tools for redoing various dialogs on demand with info preserved intact.