New feature: Skip to next trial
Added functionality where pressing “J” during a non-Hab trial ends it
and proceeds to the next trial. Deliberately does not work during hab
trials or meta-trials because J is “skip to test” in those cases, and
disabled for the first .5 seconds of a trial so that pressing J on one
trial doesn’t skip through every trial. Update for the manual coming
soon.