New habituation feature: Set first=meeting-threshold
Added a new habituation criterion mode: “Threshold”. This criterion is
set after the first X hab trials with a cumulative looking time greater
than some threshold. The ability to set the threshold and select this
mode has been added to the builder. In addition, made it so that the
trials over which the hab criterion is evaluated and the trials over
which it is set cannot ever overlap [THIS IS A FUNCTIONAL CHANGE].