Major revamp to conditions, part 1
Conditions are now stored not as indexes of stimlist but as a list of
the stimulus filenames themselves. However, both the builder and the
running script have been made backwards-compatible. In addition, and as
a result, the builder interface for conditions has been modified with
better text scaling, no random list of movie names under the condition,
and all-around better functionality. Next up, using a “pages” interface
on the conditions screen.