A kludge for PsychoPy3 compatibility
PsychoPy3 introduced an issue when you have two windows of different
sizes open on mac (distinct from retina display glitch). Basically
things need to be set up so that each window always has something in
it, and for each window, everything is drawn to it and it is flipped
before anything is drawn to the other window and it is flipped. This
adds a dummy object, the color of the background of the stimulus
window, to the back of the stimulus window to prevent that bug from
occurring.