More builder improvements, graceful failures
Changed how stimulus filenames are identified in builder to be more
reliable even if the OS for some reason returns the wrong path marker
(Windows is sometimes returning / instead of \ for some inexplicable
reason).
Added warnings to “save” and at start of run in case no trials have
been added to the study flow, to stop crashes with hard-to-interpret
error messages.