Load-and-loop, sound device, and fixing a hab bug
Two big changes:
Starting with Psychopy 3.0.5 on Mac and 3.0.6 (presumably) on PC, we
can now use the SoundDevice backend. This allows us to revert to a much
earlier design philosophy in which we load each stimulus file only once
and reset it at the end of trial, because it no longer creates horribly
audio doubling glitches (I hope)

In addition, fixed an issue where if you had no trials after your
habituation block, and the experiment ends before the last habituation
trial, the experiment would crash and fail to save.