Single-load upgrades
The old audio doubling issue returned until I realized that movie3
checker prefs.hardware, not prefs.general.
Also, if you had a trial end on movie end, the next time that movie
came up it would be blank. It turns out that this is a problem with
30fps movies because I didn’t give two frames of buffer at the end. The
system now automatically stops two frames from the end no matter what,
using dispMovie._frameInterval*2 to adjust for the frame rate of the
movie.