Fixed crashes on redos
An issue in the end experiment function would lead to a crash when
attempting to save if the last trial was not a ‘good’ trial (i.e. it
was aborted or redo was pressed after it ended).

Also some things that are stimulus-specific were reminded that they
were stimulus-specific in a continuing effort to avoid crashes when
doing no-stimulus-presentation coding.