Builder updates: Study flow and hab
A bunch of upgrades to loadFlow to allow it to render the study flow more intelligently when there are hab trials in weird places. Some remaining edge case bugs with 40+ trials, but should no longer crash.
Added pips to hab trial in study flow for sub-trials, to indicate which sub-trials go into the hab trial.
Start of bug fixes to attn-getter interface