Groundwork for hab sub-block revamp
A bunch of groundwork for better tracking of hab sub-trials. Basically, when a sub-block is created, the builder registers each trial type other than 'Hab' as 'hab_TrialType'. This is mostly out of sight of the user. During the run, the running class removes the prefix so that all of the stimulus loading and other stuff keyed by the trial type name works, but they will be saved as 'hab_TrialType' in the data, and ultimately it will help with clearing out the hab block while leaving anythign in the rest of the experiment that is of the same type intact.