Getting 'hab_' tagging working
Full implementation of tagging trials that are part of sub-blocks without breaking everything. Current behavior: Only differences are that post-Hab iterations of the trial won't be deleted at the end of habituation, and it is labeled 'hab_TRIALTYPE' in the "next trial" text and data file. Everything else, in terms of what stimuli it shows and the properties of the trial itself, are unaffected, though this actually took significant effort.