First pass at flexible habituation design
Making it so you can have these complicated hab-meta-trials and compute habituation criteria over any subset of the trials in it is extremely complicated. This is a first pass at the implementation, with I'm sure many bugs, and it hasn't been fully implemented in preflook yet. The tests also have not been updated to check the new thing, only to make sure the old way of dealing with habituation still works with a slighty tweaked data structure.

There's a lot of complex stuff under the hood. Things like using ∞ to mark the end of a meta-trial.