Making redo work with more flexible hab meta-trials, and updating tests
Attempting to implement a 'granular' redo function for hab meta-trials, so if your habituation meta-trial cares about all of these different trials it will simply subtract the looking time of that one trial, not all of them. Be sensitive to which habs are being used to calculate habituation, and even be open to the possibility that the same trial type could occur multiple times in a hab block. Tests for all of the above, which are complex!
Also fixed a couple of things in the builder that would have messed up backwards-compatibility