Changed how abort works with habituation
Because of the way the habituation calculations have been re-centered, aborttrial doesn't have to care about it provided a little more is done to recognize when something should or should not be tracked as a habituation trial.
Also refactored 'redo' to 'abort' in doTrial to more accurately reflect its function.