Catching preferential looking up to speed
One unanticipated bonus of the modifications to the habituation system that were required to make habituation over multiple sub-trials work is that the habituation calculation is based on a parallel data construct that can do all the combining as part of doTrial, and the code for checkStop is actually the same between preferential looking and single target. So, the preferential looking class is drastically reduced. The tests have been similarly pruned. All the meaningful changes are to doTrial.

Also tuned up one thing in the base class to make consistent when checkStop is called between hab meta-trials and simple 'Hab' trial type.