Experimenter screen index settings
Added a setting to change the screen index of the experimenter display, and made screen indexing flexiible enough that it should now be possible to have more than two monitors work with PyHab.

Limitation: It detects based on the number of monitors currently plugged in. It defaults to assuming there will be at least two ultimately, but if you are makign the experiment on one computer and transferring it to another, you may need to change the settings on the coputer you actually intend to run the study with.