Handling hab blocks through new block-maker interface, misc bug fixes
Since we have this lovely way of constructing blocks now, made that the primary method for constructing habituation blocks.
In the process, changed how habituation blocks work so there is no longer both a Hab *trial* and Hab *block*, only one or the other.  Changed builder interface to accomodate, and in the process modularized and expanded the delete function to be callable without the dialog.
Fixed a bug with initial rendering caused by re-ordering of code. Code much better organized across the board.
Attempting to fix an issue with determinign whether to offer the 'stimulus presentation' option. Currently would have trouble with blocks-in-blocks.