Builder fixes and launcher tweaks
Fixed builder issues where a button would show up in the wrong place, and more importantly where the list of trials that you could add stimuli to was not accounting for blocks. You cannot add stimuli directly to blocks, you can only add it to trials within blocks.

For the launcher, being smarter about checking whether there are stimuli and therefore whether to display the "stiumulus presentation on/off" switch, when dealing with blocks that can contain other blocks.