The great block implementation, hab changes
Significant revamps to testing and PyHabClass to use the generalized block structure for Hab blocks, and reserving '.' and '^' for functional use freed up a lot of previously tortured conditionals that were over-checking. Now the rules on blocks are consistent: The blocks will just hold the raw letter of their content, and are expanded only during the running of the study. The tests currently do not test blocks, and substantial work will be needed to incorporate blocks into said tests, but the basic functionality should now be there.