The big conditions builder update
The builder for conditions has been revamped.
First of all it can now handle conditions for both the order of trials within blocks and the order of stimuli within trials.
Second, it now has the ability to automatically generate conditions by permuting combinations of possible orders, and does so intelligently. You can specify if you want some trials to be immune from this, you can specify if you want it to permute every item in the list or all possible subsets.
Also put in the framework to have a set of 'base' conditions you can revert to if you randomize over participants but want to go back later.
Also rearranged the condition interface screen itself and added a 'cancel' button.

Laid groundwork for a block-level summary data file. That and getting the new conditions together will be the next, and hopefully final, piece of 0.8