Block and habituation data files
Still need to implement settings, but the under-the-hood pieces for saving block and data files have been added, with corresponding tests.
This means that each run can now save up to 6 different files. To cut down on clutter, data now has a 'verbose' sub-folder. If this folder does not exist, it will create it.
The new functions condense the data structure to one line per block or hab iteration. They are reasonably intelligent, though some caution will be needed with blocks inside of habituation blocks, which I may need to check are properly tagged in calcHabOver.