Unit test fix and sound fix
Sound fix: Will now play sounds correctly in non-stimulus-presentation modes on end of trial. The end-habituation sound still only plays once but is noticeably different from the end-trial sound.
Unit testing: Gave up on Python 2.7. Issue is not actually psychopy but one of psychopy's requirements which is no longer back-compatible with 2.7, so everything has to be tested in a 3.6 environment. This won't be an issue for anyone who has an old standalone version of 1.90.3 or earlier.