HPP setup, cleaning up everything else
Made doExperiment fully cross-compatible so it doesn't need to be rewritten, including spinning out printing data into its own function.

HPP now has experimenter display and, in theory, the key function that actually displays stimuli, though doTrial also does a lot of that lifting. The next thing HPP needs is doTrial, which is going to be the one that covers most of the differences between HPP and everything else.