New condition interface
A new, better condition-building interface based on the block-making interface (and by extension, the base UI)
Building this required expanding and improving the capabilities of loadFlow and loadTypes as well, to make them more flexible.
Also included a fix to long entries in the trial pallette in all interfaces, inserting ellipses to cap possible length.
HPP implementation for conditions incomplete, and will need additional work on the running side as well.