HPP complete
Redid how HPP works, got rid of new data structures in settings in favor of retooling stim names. HPP now basically REQUIRES you to use a conditions file, which is just the only way to manage stimuli of this complexity. However, with the new conditions builder this should not be too difficult (I hope). Also made it so conditions will be properly erased and updated, and that having a file named conditions.csv doesn't lock you into having that file forever.