Working directory fix, for now
BACK COMPATIBILITY ISSUE

Starting in PsychoPy 2020.1.0 the new runner window seems to move the working directory, which is a huge problem for PsychoPy. This manually drags the working directory back to the script location, which is how it behaved prior to 2020.1.0. Unfortunately, these extra lines of code will be needed for every existing PyHab experiment if they use 2020.1.0 or later, so it breaks back compatibility to a degree. hopefully a future change to PsychoPy renders them unnecessary.