BF: Palette pages
Fixing two bugs relating to palette pages in the builder:
1. It would not add a page correctly if you had exactly eight things and created a new block rather than a new trial.
2. It would not load multiple pages correctly when opening an existing experiment.

Right now, adding pages is a little awkward. Eventually going to make it smoother using the rounding of the trial types list, and possibly integrating it into LoadTypes (though that will need something to get you to the right page).