Slight re-tuning of "Last" hab type
Noticed that there would be an issue with "Last" hab types if the metcritwindow is greater than one. "Last" works uniquely in that if the window is greater than one, it is the only hab type that almost always incorporates trials that SET its criterion in determining whether it MET its criterion. (Max/peak w/ fixed window can theoretically do this as well but less often and it's dealt with differently in the code.)