added duration calculation toggle
Added a toggle to all trial duration calculation codeas to whether to include the last gaze-off or not. Defaults to including it, which means duration reports full trial duration (start to stop). However, when set to 0, excludes the last gaze-off if it ends the trial.

Test currently excludes last gaze-off in order to cover more conditionals in the code.