A tricky bit of logic
Problem with the auto-redo and deadline system conflicting with each other. The deadline has to effectively override the default auto-redo behavior, which is that if they haven't looked MinOn at any point when MaxOff is reached, end it then and there. The deadline gives them a set window to do that in.