edge case crash
Fixing an edge case where experiment might fail to recognize it's on the last trial if there are residual settings from a previous configuration that involved a habituation block. More generally, closes a hole in the logic that could lead to it failing to recognize the end of the experiment correctly.