hppStimScrOnly and 'gaze-off'
hppStimScrOnly now makes it so that gazes at a screen that is not the stimulus screen are counted as 'gaze-off' events for the purpose of deciding when to end the trial and for determining habituation. However, the data file is still veridical to the keys that were *actually* pressed.

Also added duration-only habituation to hab trials to HPP and PL, previously it was only set to work for hab blocks in those two modes.