each hab block only used once
Modified the builder so that a hab block could only be added to the study flow once, and hab blocks cannot be embedded in other blocks. Need to make a postdictive cleanup thing for when habituation is turned on too.

Started on some modifications to the AbortTrial function as well.