Issues with inserting hab trials
Figured out two problems with inserting a new hab trial. the big one is that the insert function could be called once per frame. The other one is that the insert could be done multiple times without acknowledging that it already happened. There also seems to be an issue with the insert occurring in not quite the right place or not returning the correct info about the new hab trial, so I tried adding a test  to catch it.