Updating demo and online edition
Demo update is straightforward. Online edition doesn't need the insert fix because it uses an older base version of the code that didn't support multiple habituation blocks and therefore didn't have the edge case in the first place, but did add the fix so it wouldn't add one per frame.