Catching up the builder for online
Brought all the changes from the in-person builder that are suitable for the online version over. The problem now is that the PyHabClass code itself needs to be updated to interpret the updated settings file. This is a breaking change, moving from 0.9 to 0.10 effectively, but it theoretically enables all the new habituation modes, block arrangement, and other features added in the most recent versions of in-person pyhab, or at least it will once PyHabClass has been similarly updated.