Minimum trial duration setting
Both builder and runner, both online and in-person. Adds a setting to the advanced trial settings that forces a trial to last a certain number of seconds regardless of gaze behavior. Could create conflicts with "end on movie end" if used incautiously. Tested out with both online and in-person versions, seems to be working. Also fixes some issues with the builder for onlien studies.