extremely sneaky calchabover bug
In hab blocks with exactly one trial, calcHabOver was being saved as a string rather than a list. This wasn't actually an issue for normal habituation, but for FixedTrialLength, it screwed up identification of the relevant trials to look for consecutive gaze-offs on. Added both a fix to the builder and a fix to PyHabClass itself that should retcon any non-list entries. Tested and working.