Updated dialogs for running study
PsychoPy 2024 introduced a change to how dialog boxes work such that they are entirely dict-based rather than list-based. As a result, every dialog that PyHab uses needs to be updated. This is the update for the dialogs that are used when running studies in-lab. These updates will need to be replicated for the online version, but also for *everything* in the builder.