Replacing MovieStim3 with MovieStim in builder
The builder needs to load the attention-getter movie files to extract their durations precisely. Unfortunately it was doing this with MovieStim3, which is now out of date. This update changes it to use MovieStim throughout. Despite the note in the documentation about duration, you do not need to start the movie to retrieve a valid duration property.