Major movie playback overhaul part 1
In trying to get consistent mid-trial AG behavior, it's become necessary to overhaul the entire movie playback system. For safety, it now extracts the first frame of the movie as an imagestim, and has a bunch of failsafes to ensure that 1) the correct frame is presented on a new trial (i.e. the first frame) and 2. the sound doesn't double up or skip in the process. The cost is that it will introduce up to 250ms of timing imprecision into the stimulus presentation itself. Reliability testing and timing solutions are still being worked on, so this is not yet release-ready.